requirements.txt
uvicorn
fastapi
torch
whisper
transformers
minio
tensorflow
sentencepiece
pydub
python-dotenv